+++ geocode_tests.py
@@ -137,6 +137,8 @@ class GeoCodeTest(unittest.TestCase):
             json.dumps({"lat": 37.794678}),
             json.dumps({"lat": "nan", "lng": -122.41143}),
             json.dumps({"lat": 91, "lng": -122.41143}),
+            json.dumps({"lat": 10**400, "lng": -122.41143}),
+            json.dumps({"lat": 37.794678, "lng": 10**400}),
         ]
 
         for cached_value in invalid_cached_values: